requirements.txt
streamlit
google-cloud-firestore
google-auth-oauthlib
firebase_admin
streamlit_option_menu
requests==2.31.0
beautifulsoup4
lxml
pandas
